Use uint8_t instead of int for SimulatorConfig::drop_percent

+++ src/io/simulator/simulator_config.hpp
@@ -20,7 +20,7 @@ namespace memgraph::io::simulator {
 using memgraph::io::Time;
 
 struct SimulatorConfig {
-  int drop_percent = 0;
+  uint8_t drop_percent = 0;
   bool perform_timeouts = false;
   bool scramble_messages = true;
   uint64_t rng_seed = 0;
